+++ src/YOKSIS/Constants/Askerlik/TeklifTuru.php
@@ -1,0 +1,12 @@
+<?php
+
+namespace Conkal\YOKSIS\Constants\Askerlik;
+
+/**
+ * AskerlikErtelemeTalep::$teklifTuru değerleri.
+ */
+class TeklifTuru
+{
+    const ERTELEME = 'E';
+    const IPTAL = 'I';
+}