+++ src/YOKSIS/Constants/Ceza/CezaMahkemeIptal.php
@@ -1,0 +1,12 @@
+<?php
+
+namespace Conkal\YOKSIS\Constants\Ceza;
+
+/**
+ * OgrenciCeza::$cezaMahkemeIptalMi değerleri.
+ */
+class CezaMahkemeIptal
+{
+    const EVET = 1;
+    const HAYIR = 2;
+}