+++ src/YOKSIS/Constants/DonemTuru.php
@@ -1,0 +1,27 @@
+<?php
+
+namespace Conkal\YOKSIS\Constants;
+
+/**
+ * Yerleştirme dönem türleri (yerlestirmeveri, toplufotografindir, vakifogrenimucreti).
+ */
+class DonemTuru
+{
+    /** Yükseköğretim Kurumları Sınavı */
+    const YKS = 'YKS';
+
+    /** YKS Ek Yerleştirme */
+    const YKS_EK = 'YKS_EK';
+
+    /** YKS Milli Sporcu */
+    const YKS_MS = 'YKS_MS';
+
+    /** Dikey Geçiş Sınavı */
+    const DGS = 'DGS';
+
+    /** DGS Ek Yerleştirme */
+    const DGS_EK = 'DGS_EK';
+
+    /** Lisans Tamamlama */
+    const LT = 'LT';
+}