+++ src/YOKSIS/REST/Entities/CezaTuru.php
@@ -1,0 +1,12 @@
+<?php
+
+namespace Conkal\YOKSIS\REST\Entities;
+
+/**
+ * Ceza türü (cezaturleri).
+ */
+class CezaTuru extends Entity
+{
+    public $kod;
+    public $ad;
+}